old_code/requirements.txt
pandas==2.2.1
pyarrow==15.0.1
geopandas==0.14.3
shapely==2.0.3
python-dotenv==1.0.1
